Rebuilt the extractData() function and getIndex() function with Obsidian MetadataCache API for MetaDataParser.
Note: The getContent function should not be rebuilt, as the frontmatterCache method in Obsidian MetadataCache will only return a map [object, Object], which could not be parsed into original markdown content. The purpose of the getContent function is to return the original text content in the YAML for replacing later. We could have trouble finding the original text content if rebuilt in the same way.